+++ grelmicro/cache/__init__.py
@@ -3,8 +3,6 @@
 from grelmicro.cache._protocol import CacheBackend
 from grelmicro.cache.cached import cached
 from grelmicro.cache.errors import CacheError, CacheSettingsValidationError
-from grelmicro.cache.memory import MemoryCacheBackend
-from grelmicro.cache.redis import RedisCacheBackend
 from grelmicro.cache.ttl import CacheInfo, TTLCache
 
 __all__ = [
@@ -12,8 +10,6 @@ __all__ = [
     "CacheError",
     "CacheInfo",
     "CacheSettingsValidationError",
-    "MemoryCacheBackend",
-    "RedisCacheBackend",
     "TTLCache",
     "cached",
 ]